fibonacci series 50 #2

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,10 +6,10 @@
  */
 int main(void)
 {
-int i, x = 1, y = 2, sum = 0;
+long int i, x = 1, y = 2, sum = 0;
 for (i = 0; i < 50; i++)
 {
-printf("%d, ", x);
+printf("%ld, ", x);
 sum = x + y;
 x = y;
 y = sum;
